+++ src/markdown2mindmap/transform.clj
@@ -3,7 +3,6 @@
             [clojure.string :as str]
             [clojure.walk :refer [prewalk]]
             [cybermonday.ir]
-            [clojure.edn :as edn]
             [markdown2mindmap.enter :as m2menter]
             [markdown2mindmap.exit :as m2mexit]
             [puget.printer :as puget]
@@ -95,7 +94,8 @@
                     FileFormatOption.)]
     (-> (SourceStringReader. puml2)
         (.outputImage out format))
-    (.close out)))
+    (.close out)
+    (printf "generated %s\n" output-file)))
 
 ;; ------------------------------------
 
@@ -132,29 +132,32 @@
          (spit output-file)))
 
 (defn md->mindmap
-  "Generates an mindmap image (with the `type` format) from a markdown file."
-  [input-file output-directory {:keys [type style with-puml]}]
-  (let [output-name (-> input-file
-                        ;; keeps only filename without extension
-                        (str/replace #"(?i)\.md" "")
-                        ;; nor path
-                        (str/replace #".*/" ""))
-        output-img (-> output-name
-                        ;; adds selected extension
-                       (str "." type)
-                       (#(io/file output-directory %)))
-        output-puml (-> output-name
-                        (str ".puml")
-                        (#(io/file output-directory %)))
-        styles (when style (slurp style))
-        puml (->> input-file
-                  slurp
-                  md->hiccup
-                  hiccup->puml
-                  (->puml2 styles))]
-    (when with-puml
-      (spit output-puml puml))
-    (create-image! output-img type puml)))
+  "Generates an mindmap image (with the `type` format) from a markdown file or dir and/or a puml file."
+  [input-file-or-dir {:keys [type style with-svg svg-output-dir with-puml puml-output-dir]}]
+  (doseq [input-file (file-seq (io/file input-file-or-dir))
+          :when (str/ends-with? input-file ".mm.md")
+          :let [svg-output-directory (or svg-output-dir (.getParent input-file))
+                puml-output-directory (or puml-output-dir (.getParent input-file))]]
+    (let [;; keeps only filename without 'md' extension
+          output-name (str/replace (.getName input-file) #"(?i)\.md" "")
+          output-img (-> output-name
+                         ;; adds selected extension
+                         (str "." type)
+                         (#(io/file svg-output-directory %)))
+          output-puml (-> output-name
+                          (str ".puml")
+                          (#(io/file puml-output-directory %)))
+          styles (when style (slurp style))
+          puml (->> input-file
+                    slurp
+                    md->hiccup
+                    hiccup->puml
+                    (->puml2 styles))]
+      (when (or with-puml puml-output-dir)
+        (spit output-puml puml)
+        (printf "generated %s\n" output-puml))
+      (when (or with-svg svg-output-dir)
+        (create-image! output-img type puml)))))
 
 (defn list-all-fonts
   "Creates an SVG image listing all fonts available on the system."