+++ test/markdown2mindmap/core_test.clj
@@ -15,5 +15,6 @@
     (testing "->puml"
       (is (=  (mth/slurp-puml n)
               (mth/hiccup-file->puml n))))
-    (sut/md->mindmap (mth/format-it "input" n "md") "output" {:type "svg" :style "resources/custom.css"})
+    (sut/md->mindmap (mth/format-it "input" n "md")
+                     {:output-dir "output" :type "svg" :style "resources/custom.css"})
     (infof "\n--- End of test %d ---" n)))